Map functions to topics via labels
Prior to this commit the Kakfa topic was inferred by parsing the
message as a JSON struct, but this is not the format messages will
arrive in within an existing system.

The topic is now mapped to a function by observing labels applied
to the containers via the gateway's /system/functions endpoint. Each
function can self-describe whether it is interested in a topic
by adding the label "topic=payment_received" for instance.

Signed-off-by: Alex Ellis (VMware) <alexellis2@gmail.com>