Add functions to get the current list of topics
This commit adds a Topics function to bring the TopicMap
implementation in line with that of the Kafka connector. The Topics
function was added to kafka-connector after the connector-sdk was
extracted from it.

It also adds a Topics function to Controller that returns the list
of topics in the TopicMap. This function is added so that connector
implementations can get the list of current topics without having to
knew about the internal structure of Controller.

Signed-off-by: Chad Taylor <taylor.thomas.c@gmail.com>