Enable multiple topics per function
The function’s topic annotation value is a CSV.  The service map should be 1..1 topic to 1..* functions. The connecter SDK currently makes the serviceMap key from the entire annotation value, which means where a CSV has been used in the function annotation value then the servicemap key will be the full CSV.

This change introduces breaking of the topic annotation CSV using comma as a delimiter so that a function can be mapped to multiple topics.

Signed-off-by: Richard Gee <richard@technologee.co.uk>