Consistent use of log.Fatal instead of panic
As discuss with @alexellis in #6 the use of log.Fatal is encouraged in
specific cases. However, `GetCredentials` in `types/credentials.go` uses
panic and misses the properly formated log statement.

This PR:

- replaces `panic` with `log.Fatalln`
- adds a doc comment for the exported function
- syncs Godeps as `dep status` even on master throws `Gopkg.lock is out
of sync with imports and/or Gopkg.toml`

The PR was tested via invoking `./tester` against the most recent
faas-netes after running `make`.

Signed-off-by: Michael Gasch <mgasch@vmware.com>